rename crate to tx-prover, use main.rs instead of bin.rs, remote lib.rs from crate
