--- bin/miden-tx-prover/src/lib.rs
@@ -1,4 +1,0 @@
-#[rustfmt::skip]
-pub mod domain;
-pub mod server;
-pub use server::generated::api::{ProveTransactionRequest, ProveTransactionResponse};
