+++ bin/tx-prover/Cargo.toml
@@ -3,7 +3,6 @@ name = "miden-tx-prover"
 version = "0.6.0"
 description = "Miden rollup transaction prover"
 readme = "README.md"
-categories = ["no-std"]
 keywords = ["miden", "transaction"]
 license.workspace = true
 authors.workspace = true
@@ -11,10 +10,6 @@ homepage.workspace = true
 repository.workspace = true
 rust-version.workspace = true
 edition.workspace = true
-
-[[bin]]
-name = "miden-tx-prover-service"
-path = "src/bin.rs"
 
 [features]
 default = ["std"]
